requirements.txt
google-cloud-aiplatform
langchain-text-splitters
pymongo[srv]
pypdf
python-dotenv
